Contact information styles. Combine main resume layouts into reusable component.
