+++ jest.config.js
@@ -1,4 +1,3 @@
-// jest.config.js
 const nextJest = require('next/jest');
 
 const createJestConfig = nextJest({
