+++ package.json
@@ -47,7 +47,7 @@
     "polished": "4.1.4",
     "prismic-javascript": "3.0.2",
     "prismic-reactjs": "1.3.4",
-    "puppeteer-core": "13.3.2",
+    "puppeteer-core": "13.4.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-transition-group": "4.4.2"
@@ -68,7 +68,7 @@
     "lint-staged": "12.3.4",
     "next-router-mock": "0.6.5",
     "prettier": "2.5.1",
-    "puppeteer": "13.3.2",
+    "puppeteer": "13.4.0",
     "ts-jest": "27.1.3",
     "typescript": "4.5.5"
   }
