+++ src/pages/private/[secret]/index.tsx
@@ -1,26 +1,8 @@
-import {
-  faBriefcase,
-  faGraduationCap,
-} from '@fortawesome/free-solid-svg-icons';
-import indefinite from 'indefinite';
-import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
-import { useRouter } from 'next/router';
+import { GetServerSideProps } from 'next';
 import React from 'react';
+import { ResumePageProps } from '../..';
 import { getCMSIntegration } from '../../../cms';
-import AboutMe from '../../../components/Articles/AboutMe';
-import ContactInformation from '../../../components/Articles/ContactInformation';
-import HobbiesAndInterests from '../../../components/Articles/HobbiesAndInterests';
-import EducationItem from '../../../components/EducationItem/EducationItem';
-import Footer from '../../../components/Footer/Footer';
-import Header from '../../../components/Header/Header';
-import PageHead from '../../../components/PageHead';
-import ProfessionalItem from '../../../components/ProfessionalItem/ProfessionalItem';
-import Section from '../../../components/Section/Section';
-import SectionHeader from '../../../components/SectionHeader/SectionHeader';
-import Skills from '../../../components/Skills/Skills';
-import { formatDate, getFullName } from '../../../helpers/utils';
-import Column from '../../../strum-design-system/components/Layout/Column';
-import Row from '../../../strum-design-system/components/Layout/Row';
+import ResumeLayout from '../../../components/Layouts/ResumeLayout';
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
   const {
@@ -31,7 +13,7 @@ export const getServerSideProps: GetServerSideProps = async (ctx) => {
   if (secret !== privateKey) {
     res.writeHead(401);
     res.end('Not authorized');
-    return { props: {} };
+    return null;
   }
 
   const CMS = getCMSIntegration();
@@ -54,102 +36,8 @@ export const getServerSideProps: GetServerSideProps = async (ctx) => {
   };
 };
 
-type ResumePageProps = InferGetServerSidePropsType<typeof getServerSideProps>;
-
 const ResumePage: React.FC<ResumePageProps> = (props) => {
-  const {
-    educationalExperiences,
-    links,
-    personalInformation,
-    privateInformation,
-    professionalExperiences,
-    skills,
-  } = props;
-  const { query } = useRouter();
-  const secret = query.secret as string;
-  const fullName = getFullName(personalInformation);
-  const jobTitle = indefinite(personalInformation.job_title);
-  const CMS = getCMSIntegration();
-
-  return (
-    <>
-      <PageHead
-        description={`Professional résumé for ${fullName}, ${jobTitle} living in ${personalInformation.location}.`}
-        personalInformation={personalInformation}
-        title={`Résumé | ${fullName} | ${personalInformation.location}`}
-      />
-
-      <Header
-        secret={secret}
-        subtitle={personalInformation.job_title}
-        title={fullName}
-      />
-
-      <Section color="standard">
-        <Row>
-          <Column>
-            <AboutMe personalInformation={personalInformation} />
-          </Column>
-          <Column>
-            <ContactInformation
-              personalInformation={personalInformation}
-              privateInformation={privateInformation}
-            />
-          </Column>
-        </Row>
-
-        <Skills skills={skills} />
-      </Section>
-
-      <Section color="alternate">
-        <SectionHeader icon={faBriefcase} text="Professional Experience" />
-        {professionalExperiences.map((experience) => (
-          <ProfessionalItem
-            end_date={
-              experience.end_date
-                ? formatDate(CMS.parseDate(experience.end_date))
-                : null
-            }
-            id={experience.id}
-            is_current={experience.is_current}
-            key={experience.id}
-            organization_name={experience.organization_name}
-            position_description={
-              <CMS.RichTextComponent
-                richText={experience.position_description}
-              />
-            }
-            position_title={experience.position_title}
-            start_date={formatDate(CMS.parseDate(experience.start_date))}
-          />
-        ))}
-      </Section>
-
-      <Section color="standard">
-        <SectionHeader icon={faGraduationCap} text="Education" />
-        {educationalExperiences.map((experience) => (
-          <EducationItem
-            achievement_description={
-              <CMS.RichTextComponent
-                richText={experience.achievement_description}
-              />
-            }
-            achievement_title={experience.achievement_title}
-            id={experience.id}
-            key={experience.id}
-            organization_name={experience.organization_name}
-            year={experience.year}
-          />
-        ))}
-      </Section>
-
-      <Section color="alternate">
-        <HobbiesAndInterests personalInformation={personalInformation} />
-      </Section>
-
-      <Footer personalInformation={personalInformation} links={links} />
-    </>
-  );
+  return <ResumeLayout {...props} />;
 };
 
 export default ResumePage;
